+++ examples/node-javascript/index.js
@@ -13,184 +13,89 @@ async function run() {
     analytics: client,
   });
 
-  RudderTyperAnalytics.productsSearched({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.sampleEvent1({
+    anonymousId: 'sample_anonymous_id',
     properties: {
-      query: 'iPhone',
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.productClicked({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      product_id: 'iphone_12',
-      sku: 'iphone_12_sku',
-      category: 'smartphones',
-      name: 'iPhone 12',
-      brand: 'Apple',
-      price: 999.99,
-      image_url: 'https://example.com/iphone_12.jpg',
-      coupon: 'SUMMER_SALE',
-      position: 1,
-      variant: 'red',
-    },
-  });
-
-  RudderTyperAnalytics.productViewed({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      product_id: 'iphone_12',
-      sku: 'iphone_12_sku',
-      category: 'smartphones',
-      name: 'iPhone 12',
-      brand: 'Apple',
-      price: 999.99,
-      image_url: 'https://example.com/iphone_12.jpg',
-      coupon: 'SUMMER_SALE',
-      url: 'https://example.com/iphone_12',
-      currency: 'USD',
-    },
-  });
-
-  RudderTyperAnalytics.productAdded({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      product_id: 'iphone_12',
-      sku: 'iphone_12_sku',
-      category: 'smartphones',
-      name: 'iPhone 12',
-      brand: 'Apple',
-      price: 999.99,
-      image_url: 'https://example.com/iphone_12.jpg',
-      coupon: 'SUMMER_SALE',
-      url: 'https://example.com/iphone_12',
-      currency: 'USD',
-      quantity: 1,
-    },
-  });
-
-  RudderTyperAnalytics.cartViewed({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      cart_id: 'cart_123',
-      products: [
-        {
-          product_id: 'iphone_12',
-          sku: 'iphone_12_sku',
-          category: 'smartphones',
-          name: 'iPhone 12',
-          brand: 'Apple',
-          price: 999.99,
-          image_url: 'https://example.com/iphone_12.jpg',
-          coupon: 'SUMMER_SALE',
-          url: 'https://example.com/iphone_12',
-          currency: 'USD',
-          quantity: 1,
+  RudderTyperAnalytics.sampleEvent1(
+    {
+      userId: 'sample_user_id',
+      anonymousId: 'sample_anonymous_id',
+      properties: {
+        'Sample property 1': 'Sample value 1',
+      },
+      context: {
+        app: {
+          name: 'RudderStack',
+          version: '1.0.0',
+          build: '100',
         },
-      ],
-    },
-  });
-
-  RudderTyperAnalytics.couponApplied({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      coupon_id: 'SUMMER_SALE',
-      cart_id: 'cart_123',
-      coupon_name: 'Summer Sale',
-      discount: 50,
-      order_id: 'order_123',
-    },
-  });
-
-  RudderTyperAnalytics.checkoutStarted({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      affiliation: 'Google Store',
-      coupon: 'SUMMER_SALE',
-      currency: 'USD',
-      discount: 50,
-      order_id: 'order_123',
-      products: [
-        {
-          product_id: 'iphone_12',
-          sku: 'iphone_12_sku',
-          category: 'smartphones',
-          name: 'iPhone 12',
-          brand: 'Apple',
-          price: 999.99,
-          image_url: 'https://example.com/iphone_12.jpg',
-          coupon: 'SUMMER_SALE',
-          url: 'https://example.com/iphone_12',
-          currency: 'USD',
-          quantity: 1,
+        device: {
+          id: 'device_id',
+          manufacturer: 'Samsung',
+          model: 'Galaxy S20',
+          name: 'Samsung Galaxy S20',
+          type: 'Android',
         },
-      ],
-      revenue: 949.99,
-      shipping: 10,
-      tax: 5,
-      value: 964.99,
+        randomKey1: 'RandomValue',
+      },
+      integrations: {
+        All: true,
+        Amplitude: false,
+      },
+      timestamp: new Date(),
+      extraKey1: 'value1',
+    },
+    () => {
+      console.log('callback called from sampleEvent1 event');
+    },
+  );
+
+  RudderTyperAnalytics.page({
+    userId: 'sample_user_id',
+    name: 'Sample Page name',
+    properties: {
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.paymentInfoEntered({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.page({
+    userId: 'sample_user_id',
     properties: {
-      checkout_id: 'checkout_123',
-      order_id: 'order_123',
-      payment_method: 'Credit Card',
-      shipping_method: 'UPS Ground',
-      step: 2,
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.checkoutStepCompleted({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.screen({
+    userId: 'sample_user_id',
+    name: 'Sample Screen name',
     properties: {
-      checkout_id: 'checkout_123',
-      payment_method: 'Credit Card',
-      shipping_method: 'UPS Ground',
-      step: 3,
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.orderCompleted({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.screen({
+    userId: 'sample_user_id',
     properties: {
-      affiliation: 'Google Store',
-      coupon: 'SUMMER_SALE',
-      currency: 'USD',
-      discount: 50,
-      order_id: 'order_123',
-      checkout_id: 'checkout_123',
-      products: [
-        {
-          product_id: 'iphone_12',
-          sku: 'iphone_12_sku',
-          category: 'smartphones',
-          name: 'iPhone 12',
-          brand: 'Apple',
-          price: 999.99,
-          image_url: 'https://example.com/iphone_12.jpg',
-          coupon: 'SUMMER_SALE',
-          url: 'https://example.com/iphone_12',
-          currency: 'USD',
-          quantity: 1,
-        },
-      ],
-      revenue: 949.99,
-      shipping: 10,
-      tax: 5,
-      value: 964.99,
+      'Sample property 1': 'Sample value 1',
+    },
+  });
+
+  RudderTyperAnalytics.identify({
+    userId: 'sample_user_id',
+    traits: {
+      'Sample property 1': 'Sample value 1',
+    },
+  });
+
+  RudderTyperAnalytics.group({
+    userId: 'sample_user_id',
+    groupId: 'sample_group_id',
+    traits: {
+      'Sample property 1': 'Sample value 1',
     },
   });