+++ examples/node-typescript/index.ts
@@ -13,29 +13,19 @@ async function run() {
     analytics: client,
   });
 
-  RudderTyperAnalytics.productsSearched({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.sampleEvent1({
+    anonymousId: 'sample_anonymous_id',
     properties: {
-      query: 'iPhone',
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.productClicked(
+  RudderTyperAnalytics.sampleEvent1(
     {
-      userId: 'temp_user_id',
-      anonymousId: 'temp_anonymous_id',
+      userId: 'sample_user_id',
+      anonymousId: 'sample_anonymous_id',
       properties: {
-        product_id: 'iphone_12',
-        sku: 'iphone_12_sku',
-        category: 'smartphones',
-        name: 'iPhone 12',
-        brand: 'Apple',
-        price: 999.99,
-        image_url: 'https://example.com/iphone_12.jpg',
-        coupon: 'SUMMER_SALE',
-        position: 1,
-        variant: 'red',
+        'Sample property 1': 'Sample value 1',
       },
       context: {
         app: {
@@ -60,163 +50,52 @@ async function run() {
       extraKey1: 'value1',
     },
     () => {
-      console.log('callback called from productClicked event');
+      console.log('callback called from sampleEvent1 event');
     },
   );
 
-  RudderTyperAnalytics.productViewed({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.page({
+    userId: 'sample_user_id',
+    name: 'Sample Page name',
     properties: {
-      product_id: 'iphone_12',
-      sku: 'iphone_12_sku',
-      category: 'smartphones',
-      name: 'iPhone 12',
-      brand: 'Apple',
-      price: 999.99,
-      image_url: 'https://example.com/iphone_12.jpg',
-      coupon: 'SUMMER_SALE',
-      url: 'https://example.com/iphone_12',
-      currency: 'USD',
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.productAdded({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.page({
+    userId: 'sample_user_id',
     properties: {
-      product_id: 'iphone_12',
-      sku: 'iphone_12_sku',
-      category: 'smartphones',
-      name: 'iPhone 12',
-      brand: 'Apple',
-      price: 999.99,
-      image_url: 'https://example.com/iphone_12.jpg',
-      coupon: 'SUMMER_SALE',
-      url: 'https://example.com/iphone_12',
-      currency: 'USD',
-      quantity: 1,
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.cartViewed({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.screen({
+    userId: 'sample_user_id',
+    name: 'Sample Screen name',
     properties: {
-      cart_id: 'cart_123',
-      products: [
-        {
-          product_id: 'iphone_12',
-          sku: 'iphone_12_sku',
-          category: 'smartphones',
-          name: 'iPhone 12',
-          brand: 'Apple',
-          price: 999.99,
-          image_url: 'https://example.com/iphone_12.jpg',
-          coupon: 'SUMMER_SALE',
-          url: 'https://example.com/iphone_12',
-          currency: 'USD',
-          quantity: 1,
-        },
-      ],
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.couponApplied({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
+  RudderTyperAnalytics.screen({
+    userId: 'sample_user_id',
     properties: {
-      coupon_id: 'SUMMER_SALE',
-      cart_id: 'cart_123',
-      coupon_name: 'Summer Sale',
-      discount: 50,
-      order_id: 'order_123',
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.checkoutStarted({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      affiliation: 'Google Store',
-      coupon: 'SUMMER_SALE',
-      currency: 'USD',
-      discount: 50,
-      order_id: 'order_123',
-      products: [
-        {
-          product_id: 'iphone_12',
-          sku: 'iphone_12_sku',
-          category: 'smartphones',
-          name: 'iPhone 12',
-          brand: 'Apple',
-          price: 999.99,
-          image_url: 'https://example.com/iphone_12.jpg',
-          coupon: 'SUMMER_SALE',
-          url: 'https://example.com/iphone_12',
-          currency: 'USD',
-          quantity: 1,
-        },
-      ],
-      revenue: 949.99,
-      shipping: 10,
-      tax: 5,
-      value: 964.99,
+  RudderTyperAnalytics.identify({
+    userId: 'sample_user_id',
+    traits: {
+      'Sample property 1': 'Sample value 1',
     },
   });
 
-  RudderTyperAnalytics.paymentInfoEntered({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      checkout_id: 'checkout_123',
-      order_id: 'order_123',
-      payment_method: 'Credit Card',
-      shipping_method: 'UPS Ground',
-      step: 2,
-    },
-  });
-
-  RudderTyperAnalytics.checkoutStepCompleted({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      checkout_id: 'checkout_123',
-      payment_method: 'Credit Card',
-      shipping_method: 'UPS Ground',
-      step: 3,
-    },
-  });
-
-  RudderTyperAnalytics.orderCompleted({
-    userId: 'temp_user_id',
-    anonymousId: 'temp_anonymous_id',
-    properties: {
-      affiliation: 'Google Store',
-      coupon: 'SUMMER_SALE',
-      currency: 'USD',
-      discount: 50,
-      order_id: 'order_123',
-      checkout_id: 'checkout_123',
-      products: [
-        {
-          product_id: 'iphone_12',
-          sku: 'iphone_12_sku',
-          category: 'smartphones',
-          name: 'iPhone 12',
-          brand: 'Apple',
-          price: 999.99,
-          image_url: 'https://example.com/iphone_12.jpg',
-          coupon: 'SUMMER_SALE',
-          url: 'https://example.com/iphone_12',
-          currency: 'USD',
-          quantity: 1,
-        },
-      ],
-      revenue: 949.99,
-      shipping: 10,
-      tax: 5,
-      value: 964.99,
+  RudderTyperAnalytics.group({
+    userId: 'sample_user_id',
+    groupId: 'sample_group_id',
+    traits: {
+      'Sample property 1': 'Sample value 1',
     },
   });