+++ src/cli/api/api.ts
@@ -4,6 +4,8 @@ import { wrapError, isWrappedError } from '../commands/error.js';
 import { sanitizeTrackingPlan } from './trackingplans.js';
 import lodash from 'lodash';
 import { APIError } from '../types.js';
+import { EventType } from 'src/generators/gen.js';
+import packageJson from '../../../package.json' assert { type: 'json' };
 
 const { set } = lodash;
 
@@ -15,6 +17,7 @@ export namespace RudderAPI {
   export type GetTrackingPlanEventsRulesResponse = {
     name: string;
     description?: string;
+    eventType: EventType;
     rules: JSONSchema7;
   };
 
@@ -65,6 +68,7 @@ export namespace RudderAPI {
   export type RuleMetadata = {
     name: string;
     description?: string;
+    eventType: EventType;
     rules: JSONSchema7;
   };
 
@@ -110,21 +114,20 @@ export async function fetchTrackingPlan(options: {
       options.email,
     );
     if (eventsResponse) {
-      const eventsRulesResponsePromise = eventsResponse.data
-        .filter((ev) => ev.eventType === 'track')
-        .map(async (ev) => {
-          const url = `tracking-plans/${options.id}/events/${ev.id}`;
-          const eventsRulesResponse = await apiGet<RudderAPI.GetTrackingPlanEventsRulesResponse>(
-            url,
-            options.token,
-            options.email,
-          );
-          return {
-            name: eventsRulesResponse.name,
-            description: eventsRulesResponse.description,
-            rules: eventsRulesResponse.rules,
-          };
-        });
+      const eventsRulesResponsePromise = eventsResponse.data.map(async (ev) => {
+        const url = `tracking-plans/${options.id}/events/${ev.id}`;
+        const eventsRulesResponse = await apiGet<RudderAPI.GetTrackingPlanEventsRulesResponse>(
+          url,
+          options.token,
+          options.email,
+        );
+        return {
+          name: eventsRulesResponse.name,
+          description: eventsRulesResponse.description,
+          eventType: eventsRulesResponse.eventType,
+          rules: eventsRulesResponse.rules,
+        };
+      });
       const eventsRulesResponse: RudderAPI.RuleMetadata[] = await Promise.all(
         eventsRulesResponsePromise,
       );
@@ -234,6 +237,7 @@ async function apiGet<T>(url: string, token: string, email: string): Promise<T> 
           url === 'workspace' || url.includes('trackingplans')
             ? `Basic ${Buffer.from(email + ':' + token).toString('base64')}`
             : 'Bearer ' + token,
+        'User-Agent': `RudderTyper/${packageJson.version} Node/${process.version} Platform/${process.platform}`,
       },
       timeout: {
         request: 10000, //ms